add needed svgs
