--- src/components/ServiceCard.tsx
@@ -1,34 +1,0 @@
-import Image from "next/image";
-import Link from "next/link";
-import { FC } from "react";
-
-type ServiceCardProps = {
-  service: {
-    imageLink: string;
-    title: string;
-    content: string;
-  };
-};
-
-const ServiceCard: FC<ServiceCardProps> = ({ service }) => {
-  return (
-    <div className="flex flex-col items-center justify-between gap-10 rounded-lg shadow-lg p-4 bg-white dark:bg-inherit dark:border-2 border-colorPrimary text-center min-w-['299px'] hover:-translate-y-4 transition-all duration-300 ">
-      <Image
-        src={service.imageLink}
-        alt="an illustration of a health of personnel teaching a group of people"
-        width={94}
-        height={96}
-      />
-      <h2 className="text-xl font-bold">{service.title}</h2>
-      <p>{service.content}</p>
-      <Link
-        href="/services"
-        className="capitalize border  border-colorPrimary rounded py-2 px-4 text-colorPrimary hover:bg-green-600 hover:text-white transition-all duration-300"
-      >
-        Learn More
-      </Link>
-    </div>
-  );
-};
-
-export default ServiceCard;
